feat : home with weather
